requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# LaTeX and PDF processing
Pillow>=10.0.0

# CLI and utilities
click>=8.0.0
rich>=13.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
